app/Python/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.12.1
Django==4.1.3
djangorestframework==3.12.4
fonttools==4.43.1
idna==3.4
importlib-resources==6.1.0
inline==0.0.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
zipp==3.17.0
